Move `SigOptImagenetExperiment` to `common_experiments.py`.
